+++ CHANGELOG.md
@@ -8,8 +8,13 @@
 0.1.1
 - Add shortcut `Ctrl+Shift+T` to print `type(variable)`, also you can change `type()` to other built-in function in extension settings.
 
-## [2022/2/14]
+## [2022/2/15]
 0.1.2
-- Downward compatible to vscode version 1.60.0.
+- Downward compatible to vscode version 1.63.0.
 - When insert simply `print()`, if current line is not empty, insert at next line, or just insert in current line.
 - Add introduction video link.
+
+## [2022/2/16]
+0.1.3
+- Able to output colored text in terminal by using python built-in package: `termcolor`.
+- Add the ability to automatically recognize variable before `+=`.